feito implementação de tenant

+++ app/Http/Controllers/Admin/PlanoController.php
@@ -42,9 +42,13 @@ class PlanoController extends Controller
     }
     public function destroy($url){
 
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with(['details'])->where('url', $url)->first();
         if (!$plan) {
             return redirect()->back();
+        }
+
+        if($plan->details->count() > 0){
+            return redirect()->back()->with('error', 'Existem detalhes neste plano, portanto não pode ser eliminado');
         }
         $plan->delete();
         return redirect()->route('plans.index');
@@ -70,11 +74,17 @@ class PlanoController extends Controller
         $plan->update($data);
         return redirect()->route('plans.index');
     }
+    public function profiles($planId){
+
+        $plan = $this->repository->where('id', $planId)->first();
+        if (!$plan) {
+            return redirect()->back();
+        }
+
+    }
     public function search(Request $request){
 
         $filters = $request->except('_token');
-
-        
         $plans = $this->repository->search($request->filter);
 
         return view('admin.pages.plans.index',[
